# backend/docs/openapi.yml
openapi: 3.0.3
info:
  title: ShelterFlex Backend API
  description: API for ShelterFlex backend service
  version: 0.1.0
  contact:
    name: ShelterFlex Team
  license:
    name: MIT
servers:
  - url: http://localhost:4000
    description: Development server
  - url: https://api.shelterflex.io
    description: Production server
paths:
...
  /api/wallet/ngn/withdraw/initiate:
    post:
      summary: Initiate NGN withdrawal
      description: Initiates a withdrawal request and moves funds from available to held.
      operationId: initiateNgnWithdrawal
      tags:
        - NGN Wallet
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WithdrawalRequest'
      responses:
        '201':
          description: Withdrawal initiated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WithdrawalResponse'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        default:
          $ref: '#/components/responses/Error'
  /api/wallet/ngn/withdrawals:
    get:
      summary: List NGN withdrawals
      description: Returns the authenticated user's NGN withdrawal history.
      operationId: listNgnWithdrawals
      tags:
        - NGN Wallet
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
        - name: cursor
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Withdrawal list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WithdrawalHistoryResponse'
        default:
          $ref: '#/components/responses/Error'
...
  /api/wallet/ngn/topup/initiate:
    post:
      summary: Initiate NGN wallet top-up
      description: Creates a payment intent for NGN wallet top-up. Returns PSP redirect URL or bank transfer details.
      operationId: initiateNgnWalletTopup
      tags:
        - NGN Wallet
      security:
        - bearerAuth: []
      parameters:
        - name: Idempotency-Key
          in: header
          required: false
          schema:
            type: string
          description: Optional idempotency key for safe retries
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - amountNgn
                - rail
              properties:
                amountNgn:
                  type: number
                  minimum: 100
                  maximum: 5000000
                rail:
                  type: string
                  enum: [paystack, flutterwave, bank_transfer, manual_admin]
      responses:
        '201':
          description: Top-up initiated
          content:
            application/json:
              schema:
                type: object
                required:
                  - success
                  - depositId
                  - externalRefSource
                  - externalRef
                properties:
                  success:
                    type: boolean
                    example: true
                  depositId:
                    type: string
                    format: uuid
                  externalRefSource:
                    type: string
                    example: paystack
                  externalRef:
                    type: string
                    example: pi_abc123
                  redirectUrl:
                    type: string
                    nullable: true
                  bankDetails:
                    type: object
                    nullable: true
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
        default:
          $ref: '#/components/responses/Error'
...
components:
  schemas:
...
    BankAccountDetails:
      type: object
      required:
        - accountNumber
        - accountName
        - bankName
      properties:
        accountNumber:
          type: string
          minLength: 10
          maxLength: 10
        accountName:
          type: string
        bankName:
          type: string
    WithdrawalRequest:
      type: object
      required:
        - amountNgn
      properties:
        amountNgn:
          type: number
          minimum: 100
          maximum: 1000000
        bankAccountRef:
          type: string
          description: Reference to a saved bank account
        bankAccount:
          $ref: '#/components/schemas/BankAccountDetails'
    WithdrawalResponse:
      type: object
      required:
        - success
        - id
        - amountNgn
        - status
        - bankAccount
        - reference
        - createdAt
        - processedAt
        - failureReason
      properties:
        success:
          type: boolean
        id:
          type: string
        amountNgn:
          type: number
        status:
          type: string
          enum: [pending, approved, rejected, confirmed, failed, reversed]
        bankAccount:
          $ref: '#/components/schemas/BankAccountDetails'
        reference:
          type: string
        createdAt:
          type: string
        processedAt:
          type: string
          nullable: true
        failureReason:
          type: string
          nullable: true
    WithdrawalHistoryResponse:
      type: object
      required:
        - success
        - entries
        - nextCursor
      properties:
        success:
          type: boolean
        entries:
          type: array
          items:
            $ref: '#/components/schemas/WithdrawalResponse'
        nextCursor:
          type: string
          nullable: true
...
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              enum:
                - VALIDATION_ERROR
                - UNAUTHORIZED
                - FORBIDDEN
                - TOO_MANY_REQUESTS
                - NOT_FOUND
                - CONFLICT
                - LISTING_ALREADY_RENTED
                - SOROBAN_ERROR
                - INTERNAL_ERROR
              example: NOT_FOUND
            message:
              type: string
              example: "Resource not found"
            details:
              type: object
              description: Optional structured context for the error
              additionalProperties: true
...
  responses:
    Error:
      description: Unexpected error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: token
...